Created the Card class
I still can't get Eclipse to work so I created the Card class using
Notepad.  Unfortunately if it's buggy I don't have a way to test it
right now.  =(
However It's a pretty simple class so I'm hoping it will work.  For both
of the data members I created enumerated types.  In order to access the
data members as Strings I added methods that allow you to grab the
card's Face and Suit as Strings.  The card's data members are only
protected so if you did want to just access it from other classes within
the same package it should be possible.